+++ example/index.ts
@@ -2,9 +2,12 @@ import { Raikiri } from '../src'
 
 const router = new Raikiri()
 
-router.add('GET', '/users/:userId', '/users/:userId')
-router.add('GET', '/game', '/game')
-router.add('GET', '/game/:gameId/state', '/game/:gameId/state')
-router.add('GET', '/game/:gameId', '/game/:gameId')
+// router.add('GET', '/api/search/:term', '/api/search/:term')
+router.add('GET', '/api/abc/view/:id', '/api/abc/view/:id')
+router.add(
+    'GET',
+    '/api/abc/:type',
+    '/api/abc/:type'
+)
 
-console.log('GOT', router.match('GET', '/game/1/state'))
+console.log('GOT', router.match('GET', '/api/abc/type'))